# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jubilujah Internal API",
    "version": "1.0.0",
    "summary": "Editorial, pipeline, ratings, awards, radio, and admin surface for Jubilujah.",
    "description": "OpenAPI 3.1 contract implementing every endpoint enumerated in Jubilujah Build Spec v1.0 §19 (API Surface — Initial). Authentication is delegated to JubileeInspire SSO (Build Spec §5); the resulting session cookie (`session_id`, HttpOnly/Secure/SameSite=Strict) authorizes all calls. All mutating requests require a CSRF token header (`X-CSRF-Token`). Role-based gates per §5: minimum `content_editor` for read; `production_manager` for pipeline transitions; `radio_producer` for playlist mutations; `admin` for `/admin/*`. Domain schemas live at `/schemas/v1/*.schema.json` and are referenced via `$ref` from this document.",
    "contact": {
      "name": "Gabriel Ungureanu — Stewardship",
      "url": "https://jubilujah.com"
    },
    "license": {
      "name": "Proprietary — Jubilee Music Internal Use Only",
      "identifier": "LicenseRef-Jubilee-Internal"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3119/api",
      "description": "Local development (server.js)"
    },
    {
      "url": "https://jubilujah.com/api",
      "description": "Production stub (gated by JubileeInspire SSO)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/artists/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "get": {
        "tags": ["Catalog"],
        "operationId": "getArtist",
        "summary": "Fetch a single artist by id.",
        "responses": {
          "200": {
            "description": "Artist found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Artist" },
                "examples": {
                  "melodyInspire": { "$ref": "#/components/examples/ArtistMelody" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Catalog"],
        "operationId": "updateArtist",
        "summary": "Partial update of an artist record.",
        "parameters": [{ "$ref": "#/components/parameters/CsrfHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ArtistPatch" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated artist.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "HttpOnly + Secure + SameSite=Strict cookie issued by the backend after a successful JubileeInspire SSO exchange (Build Spec §5)."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "UUID of the resource.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "CsrfHeader": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "description": "Double-submit CSRF token issued at login (Build Spec §20).",
        "schema": { "type": "string", "minLength": 16 }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "missingSession": {
                "value": { "error": "unauthorized", "message": "Session required. Re-authenticate via JubileeInspire SSO." }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but role does not satisfy the route's minimum.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "needsAdmin": {
                "value": { "error": "forbidden", "message": "Role 'admin' required for this operation." }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource does not exist or is not visible to the caller.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "missing": { "value": { "error": "not_found", "message": "Resource not found." } }
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Request body failed Pydantic / domain validation.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" },
            "examples": {
              "hebrewArticle": {
                "value": {
                  "error": "validation_failed",
                  "message": "Hebrew article rule violation in lyrics.",
                  "details": [
                    { "field": "lyrics.text", "code": "hebrew_article_rule", "hint": "Use 'the Ruach Kodesh' or 'Ruach HaKodesh' — never 'the Ruach HaKodesh'." }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": { "$ref": "/schemas/v1/error.schema.json" },
      "ValidationError": { "$ref": "/schemas/v1/validation-error.schema.json" },
      "Artist":         { "$ref": "/schemas/v1/artist.schema.json" },
...
      "ArtistPatch":    { "$ref": "/schemas/v1/artist.patch.schema.json" },
...
    },
    "examples": {
...
      "ArtistMelody": {
        "summary": "Single artist response for Melody Inspire.",
        "value": {
          "id": "0c0a1d2e-1111-4aaa-9000-000000000001",
          "display_name": "Melody Inspire",
          "slug": "melody-inspire",
          "grouping": "inspire_family",
          "five_fold": ["pastor", "evangelist"],
          "genre_anchor": "contemporary-worship",
          "ohi_default": true,
          "bio": "Lead vocalist of the Inspire Family; pastoral-evangelistic anchor.",
          "created_at": "2025-01-15T10:00:00Z",
          "updated_at": "2026-05-30T18:42:11Z"
        }
      },
...
    }
  }
}